Unpin json-api-resource and fix associations
The update to jsonapi-resources 0.10.6 broke the loading of tubes
for pacbio library and pool resources.

This appears to be related to the addition of the configuration option
`use_related_resource_records_for_joins`

https://github.com/cerebris/jsonapi-resources/pull/1381

This is failing to function correctly in this case.

By specifying the relationship explicitly, we disable this
behaviour, and everything behaves as expected.

I still haven't quite managed to get to the bottom of
the problem. It may be we need to disable the behaviour
globally.