Add library type as first-class citizen
Previously library type has been confined to a string in the pacbio
table.

Here we promote it to a Record, that we can then link via foreign key.
We've already set up the schema for this in the last commit.

This change lets us normalize the database, but also adds context
to better validate library type, and begin linking library types to
pipelines.