Create view for TOL tubes report
This isn't ideal because the view isn't captured by Rails in the schema.rb file so it won't be applied to new installs, only those that are migrated.

The workaround seems to be to use the Scenic Views package, but:

a) This is mostly aimed at PostGRES, and supports MySQL via adapters.
b) I don't want to add a dependency to the project just for this.